+++ lib/overcommit/hook/pre_commit/brakeman.rb
@@ -2,11 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `brakeman` against any modified Ruby/Rails files.
   class Brakeman < Base
     def run
-      unless in_path?('brakeman')
-        return :warn, 'Run `gem install brakeman`'
-      end
-
-      result = execute(%w[brakeman --exit-on-warn --quiet --summary --only-files] +
+      result = execute(%W[#{executable} --exit-on-warn --quiet --summary --only-files] +
                        applicable_files)
       return :pass if result.success?