+++ lib/overcommit/hook/pre_commit/chamber_security.rb
@@ -2,11 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `chamber secure` against any modified Chamber settings files
   class ChamberSecurity < Base
     def run
-      unless in_path?('chamber')
-        return :warn, 'Run `gem install chamber`'
-      end
-
-      result = execute(%w[chamber secure --echo --files] + applicable_files)
+      result = execute(%W[#{executable} secure --echo --files] + applicable_files)
 
       return :pass if result.stdout.empty?
       [:fail, "These settings appear to need to be secured but were not: #{result.stdout}"]