+++ lib/overcommit/hook/pre_commit/coffee_lint.rb
@@ -2,11 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `coffeelint` against any modified CoffeeScript files.
   class CoffeeLint < Base
     def run
-      unless in_path?('coffeelint')
-        return :warn, 'Run `npm install -g coffeelint`'
-      end
-
-      result = execute(%w[coffeelint --quiet] + applicable_files)
+      result = execute(%W[#{executable} --quiet] + applicable_files)
       return :pass if result.success?
 
       [:fail, result.stdout]