+++ lib/overcommit/hook/pre_commit/js_hint.rb
@@ -2,11 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `jshint` against any modified JavaScript files.
   class JsHint < Base
     def run
-      unless in_path?('jshint')
-        return :warn, 'jshint not installed -- run `npm install -g jshint`'
-      end
-
-      result = execute(%w[jshint] + applicable_files)
+      result = execute([executable] + applicable_files)
       output = result.stdout
 
       return :pass if output.empty?