+++ spec/overcommit/hook/pre_commit/bundle_check_spec.rb
@@ -5,59 +5,45 @@ describe Overcommit::Hook::PreCommit::BundleCheck do
   let(:context) { double('context') }
   subject { described_class.new(config, context) }
 
-  context 'when bundler is not installed' do
-    before do
-      subject.stub(:in_path?).and_return(false)
+  context 'when Gemfile.lock is ignored' do
+    around do |example|
+      repo do
+        `touch Gemfile.lock`
+        `echo Gemfile.lock > .gitignore`
+        `git add .gitignore`
+        `git commit -m "Ignore Gemfile.lock"`
+        example.run
+      end
     end
 
-    it { should warn }
+    it { should pass }
   end
 
-  context 'when bundler is installed' do
-    before do
-      subject.stub(:in_path?).and_return(true)
+  context 'when Gemfile.lock is not ignored' do
+    let(:result) { double('result') }
+
+    around do |example|
+      repo do
+        example.run
+      end
     end
 
-    context 'when Gemfile.lock is ignored' do
-      around do |example|
-        repo do
-          `touch Gemfile.lock`
-          `echo Gemfile.lock > .gitignore`
-          `git add .gitignore`
-          `git commit -m "Ignore Gemfile.lock"`
-          example.run
-        end
-      end
+    before do
+      result.stub(:success? => success, :stdout => 'Bundler error message')
+      subject.stub(:execute).and_call_original
+      subject.stub(:execute).with(%w[bundle check]).and_return(result)
+    end
+
+    context 'and bundle check exits unsuccessfully' do
+      let(:success) { false }
+
+      it { should fail_hook }
+    end
+
+    context 'and bundle check exist successfully' do
+      let(:success) { true }
 
       it { should pass }
-    end
-
-    context 'when Gemfile.lock is not ignored' do
-      let(:result) { double('result') }
-
-      around do |example|
-        repo do
-          example.run
-        end
-      end
-
-      before do
-        result.stub(:success? => success, :stdout => 'Bundler error message')
-        subject.stub(:execute).and_call_original
-        subject.stub(:execute).with(%w[bundle check]).and_return(result)
-      end
-
-      context 'and bundle check exits unsuccessfully' do
-        let(:success) { false }
-
-        it { should fail_hook }
-      end
-
-      context 'and bundle check exist successfully' do
-        let(:success) { true }
-
-        it { should pass }
-      end
     end
   end
 end