+++ spec/overcommit/hook/pre_commit/chamber_security_spec.rb
@@ -6,16 +6,7 @@ describe Overcommit::Hook::PreCommit::ChamberSecurity do
   subject       { described_class.new(config, context) }
 
   before do
-    subject.stub(:in_path?).and_return(true)
     subject.stub(:applicable_files).and_return(['my_settings.yml'])
-  end
-
-  context 'when chamber is not installed' do
-    before do
-      subject.stub(:in_path?).and_return(false)
-    end
-
-    it { should warn }
   end
 
   context 'when chamber exits successfully' do