+++ spec/overcommit/hook/pre_commit/coffee_lint_spec.rb
@@ -6,16 +6,7 @@ describe Overcommit::Hook::PreCommit::CoffeeLint do
   subject { described_class.new(config, context) }
 
   before do
-    subject.stub(:in_path?).and_return(true)
     subject.stub(:applicable_files).and_return(%w[file1.coffee file2.coffee])
-  end
-
-  context 'when coffeelint is not installed' do
-    before do
-      subject.stub(:in_path?).and_return(false)
-    end
-
-    it { should warn }
   end
 
   context 'when coffeelint exits successfully' do