+++ spec/overcommit/hook/pre_commit/scss_lint_spec.rb
@@ -6,16 +6,7 @@ describe Overcommit::Hook::PreCommit::ScssLint do
   subject { described_class.new(config, context) }
 
   before do
-    subject.stub(:in_path?).and_return(true)
     subject.stub(:applicable_files).and_return(%w[file1.scss file2.scss])
-  end
-
-  context 'when scss-lint is not installed' do
-    before do
-      subject.stub(:in_path?).and_return(false)
-    end
-
-    it { should warn }
   end
 
   context 'when scss-lint exits successfully' do