+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * Fix bug where pre-commit hook would crash if user attempted to commit a
   broken symlink
 * Add `BrokenSymlinks` pre-commit hook which checks for broken symlinks
+* Fix Chamber integration
+* Fix 'include' path for ChamberSecurity
 
 ## 0.17.0