+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Overcommit Changelog
 
-## master (unreleased)
+## 0.18.0
 
 * Update minimum version of `image_optim` gem to 0.15.0 (breaking change in
   name of exception classes)